Extract: move shared form into components instead of pages
